+++ NQueen.Kernel/Models/SolutionUpdateDTO.cs
@@ -2,13 +2,13 @@
 
 public class SolutionUpdateDTO
 {
-    public byte BoardSize { get; set; }
+    public int BoardSize { get; set; }
 
     public SolutionMode SolutionMode { get; set; }
 
-    public byte[] QueenPositions { get; set; }
+    public int[] QueenPositions { get; set; }
 
-    public HashSet<byte[]> Solutions { get; set; }
+    public HashSet<int[]> Solutions { get; set; }
 
     public int NoOfSolution => Solutions.Count;
 }
